Дефицит графа и почти оконченный дефицит по татту бержу
