+++ TattBergeDeficite/src/main/java/org/example/DeficiteByTattBerge.java
@@ -1,12 +1,18 @@
+
 package org.example;
 
 import com.mathsystem.api.graph.model.Edge;
-import com.mathsystem.api.graph.model.Graph;
 import com.mathsystem.api.graph.model.Vertex;
+import com.mathsystem.api.graph.model.Graph;
 import com.mathsystem.domain.graph.repository.Color;
 import com.mathsystem.domain.graph.repository.GraphType;
 import com.mathsystem.domain.plugin.plugintype.GraphCharacteristic;
 
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+import java.util.UUID;
+import javax.lang.model.type.NullType;
 import java.util.*;
 
 public class DeficiteByTattBerge implements GraphCharacteristic {
@@ -20,7 +26,9 @@ public class DeficiteByTattBerge implements GraphCharacteristic {
         List<Edge> array = graph.getEdges();
         for (Edge edge : array) {
             map.get(edge.getToV()).setWeight(0);
+            map.get(edge.getToV()).setColor(Color.gray);
             map.get(edge.getFromV()).setWeight(0);
+            map.get(edge.getFromV()).setColor(Color.gray);
         }
 
         for (Edge edge : array) {
@@ -55,15 +63,18 @@ public class DeficiteByTattBerge implements GraphCharacteristic {
             // вставляем подмножество в множество
             set.add(subset);
         }
-        int def = graph.getVertexCount();
+        int def = -graph.getVertexCount();
         for (List<UUID> list : set) {
+            // количество компонент связности при удалении данного множества вершин
             int k = FindCountOfConnectivityComponents(
                     new Graph(graph.getDirectType(),
                             graph.getVertexCount(),
                             graph.getEdgeCount(),
                             graph.getVertices(),
-                            graph.getEdges()), list);// количество компонент связности при удалении данного множества вершин
-            def = Math.max(def, k-list.size());
+                            graph.getEdges()), list);
+            if (k!=-1) {
+                def = Math.max(def, k- list.size());
+            }
         }
         return def;
     }
@@ -72,15 +83,35 @@ public class DeficiteByTattBerge implements GraphCharacteristic {
         List<Edge> edgeList = graph.getEdges();
         List<Edge> newEdges = new ArrayList<>();
         Map<UUID,Vertex> newMap = new HashMap<>();
+        boolean flag;
         for (Edge edge : edgeList) {
+            flag=true;
             for (UUID uuid: set) {
-                if (uuid!=edge.getToV() || uuid!=edge.getFromV()) {
-                    newMap.put(uuid, oldMap.get(uuid));
-                    newEdges.add(edge);
+                if (uuid==edge.getToV() || uuid==edge.getFromV()) {
+                    flag=false;
+                    break;
                 }
+            }
+            if (flag) {
+                newEdges.add(edge);
+            }
+        }
+        for (UUID uuid: oldMap.keySet()) {
+            flag=true;
+            for (UUID v : set) {
+                if(v==uuid) {
+                    flag=false;
+                    break;
+                }
+            }
+            if (flag) {
+                newMap.put(uuid, oldMap.get(uuid));
             }
         }
         Graph newGraph = new Graph(graph.getDirectType(), newMap.size(), newEdges.size(), newMap, newEdges);
+        if (newGraph.getVertexCount()%2==0) {
+            return -1;
+        }
         Map<UUID, Map<UUID, Boolean>> adjacencyMatrix = new HashMap<>();
         Map<UUID, Boolean> visited = new HashMap<>();
         List<UUID> stack = new ArrayList<>();
@@ -94,16 +125,16 @@ public class DeficiteByTattBerge implements GraphCharacteristic {
 
         for (Map.Entry<UUID, Map<UUID, Boolean>> vertex : adjacencyMatrix.entrySet()) {
             if (!visited.get(vertex.getKey())) {
-                DFS(graph, vertex.getKey(), adjacencyMatrix, visited, stack);
+                currentComponent++;
+                DFS(newGraph, vertex.getKey(), adjacencyMatrix, visited, stack);
             }
         }
-
         return currentComponent;
     }
     private void DFS(Graph graph, UUID vertex, Map<UUID, Map<UUID, Boolean>> adjacencyMatrix, Map<UUID, Boolean> visited, List<UUID> stack) {
         visited.put(vertex, true);
         for (Map.Entry<UUID, Map<UUID, Boolean>> neighbour : adjacencyMatrix.entrySet()) {
-            if (adjacencyMatrix.get(vertex).get(neighbour.getKey()) && !visited.get(neighbour.getKey())) {
+            if (!visited.get(neighbour.getKey())) {
                 DFS(graph, neighbour.getKey(), adjacencyMatrix, visited, stack);
             }
         }
@@ -127,13 +158,14 @@ public class DeficiteByTattBerge implements GraphCharacteristic {
         for (Edge edge : graph.getEdges()) {
             UUID vertex1 = edge.getFromV();
             UUID vertex2 = edge.getToV();
-
-            if (graph.getVertices().get(vertex1).getColor() == Color.gray &&
-                    graph.getVertices().get(vertex2).getColor() == Color.gray) {
-
-                adjacencyMatrix.get(vertex1).put(vertex2, true);
-                if (graph.getDirectType() == GraphType.UNDIRECTED) {
-                    adjacencyMatrix.get(vertex2).put(vertex1, true);
+            if (graph.getVertices().get(vertex1) != null &&
+                    graph.getVertices().get(vertex2) != null) {
+                if (graph.getVertices().get(vertex1).getColor() == Color.gray &&
+                        graph.getVertices().get(vertex2).getColor() == Color.gray) {
+                    adjacencyMatrix.get(vertex1).put(vertex2, true);
+                    if (graph.getDirectType() == GraphType.UNDIRECTED) {
+                        adjacencyMatrix.get(vertex2).put(vertex1, true);
+                    }
                 }
             }
         }
