+++ TattBergeDeficite/src/test/resources/Berge0.txt
@@ -1,0 +1,5 @@
+UNDIRECTED 2
+ 1
+985282f5-9f05-45f5-b88f-df49bc93d0ba gray null null 159 82
+8709c807-0b79-4f5b-b4c4-3c6c781fb61f gray null null 408 138
+985282f5-9f05-45f5-b88f-df49bc93d0ba 8709c807-0b79-4f5b-b4c4-3c6c781fb61f gray null null
