+++ TattBergeDeficite/src/test/resources/Berge1.txt
@@ -1,0 +1,8 @@
+UNDIRECTED 3
+ 3
+e8e2a9ef-7c34-4185-a777-1170d546b8d0 gray null null 473 84
+0c3b7144-5ce6-46a9-b422-b42724d43c68 gray null null 385 214
+fd6218f0-63a1-4e54-a1b1-a6e129339149 gray null null 269 65
+fd6218f0-63a1-4e54-a1b1-a6e129339149 e8e2a9ef-7c34-4185-a777-1170d546b8d0 gray null null
+e8e2a9ef-7c34-4185-a777-1170d546b8d0 0c3b7144-5ce6-46a9-b422-b42724d43c68 gray null null
+0c3b7144-5ce6-46a9-b422-b42724d43c68 fd6218f0-63a1-4e54-a1b1-a6e129339149 gray null null
