+++ TattBergeDeficite/src/test/resources/Berge4.txt
@@ -1,0 +1,21 @@
+UNDIRECTED 8
+ 11
+539006d6-e935-4f57-81bd-ffb298f650e6 gray null null 249 51
+01d801c3-1c95-49ec-9ebb-b1eb1ce196bb gray null null 134 148
+4b234a53-def6-415e-9976-ee5aa0cbf38a gray null null 251 144
+b1e2b3cd-f626-4ec5-af99-e80f92f053cf gray null null 345 144
+18083481-832f-4a88-b536-9d0c0cbcc605 gray null null 132 244
+f3d154a7-5399-48e7-9d4b-3aeb38ad9742 gray null null 245 234
+4cc8278f-93c7-427e-8769-6830570ed5f2 gray null null 340 243
+fe813528-7089-4793-ad30-4664ee523377 gray null null 246 350
+539006d6-e935-4f57-81bd-ffb298f650e6 01d801c3-1c95-49ec-9ebb-b1eb1ce196bb gray null null
+539006d6-e935-4f57-81bd-ffb298f650e6 4b234a53-def6-415e-9976-ee5aa0cbf38a gray null null
+539006d6-e935-4f57-81bd-ffb298f650e6 b1e2b3cd-f626-4ec5-af99-e80f92f053cf gray null null
+01d801c3-1c95-49ec-9ebb-b1eb1ce196bb 18083481-832f-4a88-b536-9d0c0cbcc605 gray null null
+4b234a53-def6-415e-9976-ee5aa0cbf38a 18083481-832f-4a88-b536-9d0c0cbcc605 gray null null
+4b234a53-def6-415e-9976-ee5aa0cbf38a f3d154a7-5399-48e7-9d4b-3aeb38ad9742 gray null null
+b1e2b3cd-f626-4ec5-af99-e80f92f053cf 4cc8278f-93c7-427e-8769-6830570ed5f2 gray null null
+4cc8278f-93c7-427e-8769-6830570ed5f2 f3d154a7-5399-48e7-9d4b-3aeb38ad9742 gray null null
+f3d154a7-5399-48e7-9d4b-3aeb38ad9742 fe813528-7089-4793-ad30-4664ee523377 gray null null
+18083481-832f-4a88-b536-9d0c0cbcc605 fe813528-7089-4793-ad30-4664ee523377 gray null null
+4cc8278f-93c7-427e-8769-6830570ed5f2 fe813528-7089-4793-ad30-4664ee523377 gray null null
