+++ TattBergeDeficite/src/test/resources/Berge5.txt
@@ -1,0 +1,17 @@
+UNDIRECTED 5
+ 10
+5fd66824-28dd-430a-836e-ae08f2599a95 gray null null 182 151
+efb0d145-4e86-40da-9cc0-7e7c363bcee7 gray null null 292 68
+8aca00a8-9c46-465b-aa8f-860587bf299a gray null null 388 162
+a2738995-8d87-4558-b3b6-21678f800cb0 gray null null 350 300
+b1bb233c-135d-4895-9b41-6d7d78c948e9 gray null null 218 300
+b1bb233c-135d-4895-9b41-6d7d78c948e9 5fd66824-28dd-430a-836e-ae08f2599a95 gray null null
+5fd66824-28dd-430a-836e-ae08f2599a95 efb0d145-4e86-40da-9cc0-7e7c363bcee7 gray null null
+efb0d145-4e86-40da-9cc0-7e7c363bcee7 8aca00a8-9c46-465b-aa8f-860587bf299a gray null null
+8aca00a8-9c46-465b-aa8f-860587bf299a a2738995-8d87-4558-b3b6-21678f800cb0 gray null null
+a2738995-8d87-4558-b3b6-21678f800cb0 b1bb233c-135d-4895-9b41-6d7d78c948e9 gray null null
+b1bb233c-135d-4895-9b41-6d7d78c948e9 8aca00a8-9c46-465b-aa8f-860587bf299a gray null null
+b1bb233c-135d-4895-9b41-6d7d78c948e9 efb0d145-4e86-40da-9cc0-7e7c363bcee7 gray null null
+efb0d145-4e86-40da-9cc0-7e7c363bcee7 a2738995-8d87-4558-b3b6-21678f800cb0 gray null null
+8aca00a8-9c46-465b-aa8f-860587bf299a 5fd66824-28dd-430a-836e-ae08f2599a95 gray null null
+5fd66824-28dd-430a-836e-ae08f2599a95 a2738995-8d87-4558-b3b6-21678f800cb0 gray null null
