+++ TattBergeDeficite/src/test/resources/Berge6.txt
@@ -1,0 +1,102 @@
+UNDIRECTED 100
+ 0
+11111111-1111-1111-1111-111111111111 gray null null 63 55
+3fb18d9a-5d46-438d-8cd6-eb4e07e51e6c gray null null 129 55
+6a91b822-601b-470e-98ea-db68951fa8ad gray null null 174 57
+e603a072-ff0c-41fb-a4c0-8be31ab487d5 gray null null 222 57
+742ec118-df27-477f-91b7-3dd6d4bab63b gray null null 263 57
+0cd48c91-e9c7-451f-bf1d-8647165388c9 gray null null 290 57
+c2e41d38-7a55-44a6-8cd9-b752e76ad059 gray null null 332 57
+66accde8-f907-4608-bf11-e73807089e16 gray null null 356 57
+731b35cf-b71a-442d-a6af-f4fdab5ab18a gray null null 380 60
+4250d2c1-02b1-4b74-b216-87a74da5a45b gray null null 408 60
+40b197d3-9eb9-42bc-a408-accd72053d24 gray null null 435 60
+3ca47e50-d987-4169-ba21-6043e03713bd gray null null 455 62
+c75abc6f-f02a-43ef-b75a-ae74fe4d192f gray null null 481 64
+778cd427-a051-4740-b47a-ae3f09989834 gray null null 515 67
+e9650ed5-dccd-4ff2-97b9-f377cca1d05b gray null null 541 78
+1d0c01f2-9b81-44f2-a741-5e809abf3e5c gray null null 553 104
+453ea69f-33a9-4456-93ca-53e1fcb42896 gray null null 553 130
+9779bb9e-b688-4085-bf94-775bfbef8535 gray null null 548 164
+a1be7076-0342-4dae-ab15-e4385078243c gray null null 526 178
+54d3d9af-c5de-412f-b6af-03473e3ef530 gray null null 493 178
+117e4f05-d0fb-4f11-9b28-41d5af24136d gray null null 482 148
+24aa7ea5-829a-44e1-a314-0be2376d0838 gray null null 503 127
+9b2a035c-77be-4618-bee8-f6a3d22a2d25 gray null null 512 132
+8ee68980-fce3-48bf-9780-c92d308cb439 gray null null 413 166
+cb982be9-1ec4-443e-b981-9394ed7f69b7 gray null null 368 121
+85e08283-3a6a-46e2-a133-3c0bc79b52b0 gray null null 307 163
+02df0bf1-ed70-4a97-97cb-d0aa64afebf9 gray null null 370 200
+af011847-26ed-41b6-ab55-a4a8233395aa gray null null 417 94
+80679f17-a73f-4f00-b59e-7314ee2a1614 gray null null 251 103
+71e11bf1-f404-40b8-ab38-393fa0701558 gray null null 423 213
+c2d38655-9d11-42dd-b3b3-5e86585b20fb gray null null 298 237
+79b51b7e-cf0e-4288-9c95-db364328cc01 gray null null 351 155
+461aee8f-3610-4218-bbe9-cea9c80b12d2 gray null null 433 121
+0ebab7d0-1b62-4fa1-89ea-929433297209 gray null null 335 100
+b4efe0b0-e7a4-41d0-b6a6-4fb502ea8970 gray null null 241 182
+08e02472-13df-4086-a43b-19d9ae04fed7 gray null null 329 233
+7631589f-4189-48bb-bb9e-0f29b296bb1a gray null null 381 240
+841aede8-d514-4b41-a4f6-5a8a9a8f70c0 gray null null 394 129
+8576b8ff-e645-46df-94d9-5afa3f999ba6 gray null null 282 136
+b95aa0d4-1b36-4c37-ae31-d43017367344 gray null null 297 98
+54089f2b-f586-4221-8807-6ad4a215b59b gray null null 325 136
+dead1f4f-4c85-4b9b-a79e-c4d167a7c2b0 gray null null 505 96
+d6a77ee6-c6a9-4756-8d16-120b751e1d22 gray null null 455 93
+f5adc9ca-cd2a-4db7-8ac4-9096b63432d1 gray null null 473 94
+1f64c17c-58d4-4fdf-a392-892743f9b944 gray null null 461 121
+c457704d-406e-4ebb-96c8-4d7856ddfcf6 gray null null 448 147
+1a46214d-5746-45ae-bb24-2a5275dd1392 gray null null 464 174
+05e836c9-a127-4481-b2c3-899794d91c55 gray null null 433 185
+60cfd790-40ce-4109-b752-2e5b85388deb gray null null 460 203
+986d4cad-fef6-4e39-93a2-989d19f51b91 gray null null 490 200
+e87e28ee-cde0-4dce-baee-96abcde23bd0 gray null null 520 205
+384b68e8-2d45-4119-b120-36016a7a4a1f gray null null 516 231
+423ed71d-0712-497a-b28f-1816bd5ab694 gray null null 492 224
+c1334f51-c93f-4639-9f4c-e10577c1b7e0 gray null null 465 226
+e24c3431-b2a6-4016-9f35-b5a87e9c46e2 gray null null 432 236
+68cffd3f-25ec-49e1-aa43-fb792953b51b gray null null 403 252
+a32f2dd8-63b3-45d4-855d-7a77f0a5949d gray null null 417 284
+a6e37198-4fbf-428c-9696-8abc7e251808 gray null null 447 305
+f8975279-4a15-4bf1-998c-fef8f3837120 gray null null 480 294
+f99ab440-cc5a-4b08-a084-4b58ae68f408 gray null null 469 262
+ec231f1b-00b3-41e4-adfe-556a57572c05 gray null null 441 262
+4e1cd3af-cd50-421f-a84f-99f664e3f380 gray null null 487 256
+cffc15cf-f6e7-4ad9-be49-12e7e09948b7 gray null null 516 261
+7cfdfc8c-cdb8-443f-b25c-7cc4dc8df1cd gray null null 532 267
+145995cb-131f-4391-8c36-5545b98963e2 gray null null 552 276
+222c67a5-98c8-45a0-b6b2-41d337d5dc27 gray null null 582 309
+f97367c2-c7aa-4777-a495-9b1716351d47 gray null null 585 313
+98b21f53-56b1-4a51-a113-a76bdd1f9988 gray null null 588 341
+130f1900-be21-450f-926d-339eff745eb1 gray null null 555 349
+f8a6887f-6e30-49f1-bf69-f83f177cdd12 gray null null 528 316
+d78a78af-ad5a-4ae0-b78b-093ed807d4da gray null null 574 257
+86509fc1-6f93-45cf-9dd9-d1db1f4372e9 gray null null 588 219
+f2df46a2-ac95-4d7e-8eb4-b1c164d46d9d gray null null 601 184
+3535a0fc-7847-435f-af54-d62f986920c6 gray null null 604 151
+0111f7f8-e2a1-445b-9fff-4eac5e6ab894 gray null null 599 89
+d6461b2f-349f-4d2d-a181-aea5fbf2e364 gray null null 594 60
+05b474c3-063b-442f-b5b9-a4411f845669 gray null null 621 44
+41db620b-cc4a-42a6-b8f9-08176438140b gray null null 629 78
+78eec605-7a34-44ca-860c-31da2d0f1b69 gray null null 632 105
+7e6a43c7-4c7b-4af6-a51c-d206a043447c gray null null 631 142
+cfec4c67-dfaf-4faf-9b42-f7deb06a6d37 gray null null 630 186
+f8e2ebb4-8ba5-46ce-859b-f83221bf7597 gray null null 630 245
+ef85a533-3608-452f-aa09-456fef2c08d0 gray null null 630 279
+a900a032-d44e-4135-a8d8-84072d258d8d gray null null 627 331
+2a7e560f-ab9b-4fd3-823b-82eac4bdcbde gray null null 616 374
+12463ffd-b14b-4628-9330-b18eb11f2661 gray null null 579 375
+560917ca-9ea0-4fda-b142-ba8f5ccc78f7 gray null null 525 379
+0d2fac85-6e79-477f-9f52-776718f7c15d gray null null 475 359
+44586fac-f731-4af8-979a-5c5d356376bb gray null null 407 346
+8f30d2c1-d36b-477d-8660-d15aefadc9ea gray null null 363 325
+6e21b6d4-7914-4658-9b8c-f9d4317ba0c2 gray null null 324 303
+1dc11642-946f-4f66-b476-b8c892b317bb gray null null 282 285
+7af1d411-a64c-4349-9639-c9030a719177 gray null null 229 255
+53d1e8e7-3e5b-4c2e-8440-fec4aa3367ed gray null null 179 223
+96a2f4eb-a705-4ab4-83e1-d99ce84fe9d5 gray null null 133 183
+f7bc7f2c-3915-4600-97e4-f4460b3c868e gray null null 107 166
+bd032f16-2f1a-4f2e-8dad-1a699e785e01 gray null null 78 136
+288f3764-db2e-41dd-83bf-7b77a757156b gray null null 58 91
+a955de51-6211-4cb1-b481-4221f977a976 gray null null 29 67
+b9b88a2d-2b23-4220-8e01-78092a3f14fb gray null null 93 84
